pyproject.toml
[project]
name = "lelm"
version = "0.1.0"
description = "Fine-tuned LLM for NBA hot takes"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "transformers>=4.43",
    "trl>=0.9",
    "datasets>=2.20",
    "peft>=0.12",
    "bitsandbytes>=0.43",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
unsloth = [
    "unsloth[colab-new]>=2024.8",
]
